+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Automatic R packages installation (argparser and ggplot2).
+
 ### Changed
 - Separated curve calculation and fraction calculation to different functions.
 - Harmonized `melt_X.py` scripts output.